Crate problem085

Problem 85 (Counting rectangles)

Solution detail

For an m × n grid, the number of sub-rectangles is given by Tm × Tn, where Tk = k * (k + 1) / 2 is the kth triangular number.

This is because there are Tm choices for the projection of the rectangle onto the x-axis, similarly Tn choices for the projection onto the y-axis, and these projections uniquely determing the rectangle.

This means that we just need to generate the triangular numbers, up to and including the first one which exceeds the target number of sub-rectangles, and for each one, find the triangular number that pairs with it to give the product closest to the target.

Doing this for each triangular number, and keeping track of the best rectangle so far, we will arrive at the answer.

Functions

answer
Solve the problem, returning the answer as a String

first_index_greater_than
Find the index of the first entry in the (sorted!) vector which is greater than the target.

main

solve
Find the area of the rectangle with a number of sub-rectangles closest to the given target.

triangle
Calculate the nth triangular number

triangles_until
Find the triangular numbers up to and including the first one which is at least target.